% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSAvis_shiny.R
\name{BSAvis_shiny}
\alias{BSAvis_shiny}
\title{Run BSAvis}
\usage{
BSAvis_shiny(vcf.list)
}
\arguments{
\item{vcf.list}{object containing meta information and vcf data frame}
}
\description{
Run the interactive BSA visualisation tool.
}
\examples{
BSAvis_shiny(vcf.list=vcf_list)
}
