% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_deltaSNPindex.R
\name{calc_deltaSNPindex}
\alias{calc_deltaSNPindex}
\title{Calculate delta(SNP index)}
\usage{
calc_deltaSNPindex(SNPindex.windows)
}
\arguments{
\item{SNPindex.windows}{data frame obtained after running the slidingWindow()function}
}
\value{
Data frame containing start, mid and stop positions of each window in the chromosome, as well as the corresponding mean SNP-index value for each bulk and ∆(SNP-index)value.
}
\description{
This function is used to calculate the ∆(SNP-index) by subtracting the SNP-index value of the wild-type bulk from the SNP-index value of the mutant bulk. 

\deqn{mean_M_SNPindex - mean_WT_SNPindex.WT}
}
\details{
For each row in the data frame returned by slidingWindow(), corresponding to a different window, ∆(SNP-index) is calculated by subtracting the SNP-index value of the reference bulk to the SNP-index value of the mutant bulk. 
The values of ∆(SNP-index)ineach window are added to the data frame as a new column.
}
\examples{
deltaSNPindex_windows <- calc_deltaSNPindex(SNPindex.windows=SNPindex_windows) 
}
