% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SIMPLEratio.R
\name{calc_SIMPLEratio}
\alias{calc_SIMPLEratio}
\title{Calculate SIMPLE-ratio}
\usage{
calc_SIMPLEratio(vcf.df, WTbulk, Mbulk)
}
\arguments{
\item{vcf.df}{vcf data frame}

\item{WTbulk}{wild-type bulk}

\item{Mbulk}{mutant bulk}
}
\value{
data frame (SIMPLE-ratio)
}
\description{
First of all, this function creates a data frame for each bulk, which includes a column containing the SNP-index values. 
Then, rows containing InDel variants (i.e. insertions/deletions) get removed, and the data frames get joined together by chromosome (ChromKey) and position (POS).\n
Finally, the SIMPLE-ratio between the two bulks (M/WT) gets calculated as follows:
\deqn{(AD_ref.WT/(AD_ref.WT + AD_alt.WT)) - (AD_ref.M/(AD_ref.M + AD_alt.M))}\n
Reference: Wachsman et al., 2017
}
\examples{
vcf_df_SIMPLEratio <- calc_SIMPLEratio(vcf.df=vcf_df, 
WTbulk="pool_S3781_minus", 
Mbulk="pool_S3781_plus")
}
