% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SNPratio.R
\name{calc_SNPratio}
\alias{calc_SNPratio}
\title{Calculate SNP-ratio}
\usage{
calc_SNPratio(vcf.df, WTbulk, Mbulk, Variants = "SNP")
}
\arguments{
\item{vcf.df}{vcf data frame}

\item{WTbulk}{wild-type bulk}

\item{Mbulk}{mutant bulk}

\item{Variants}{variants to be considered. Default is "SNP" (allowed: "SNP" or "all")}
}
\value{
data frame (SNP-ratio)
}
\description{
This function allows to calculate the SNP-ratio between two bulks (M/WT), as follows:
\deqn{(AD_ref.WT/(AD_ref.WT + AD_alt.WT)) - (AD_ref.M/(AD_ref.M + AD_alt.M))}
Note that the user can select specific variants to consider, by setting the "Variants" parameter to "SNP" (default) or "all" (respectively referring to SNPs or InDels+SNPs).
}
\references{
Wachsman et al., 2017
}
