% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaSNPindex_plot.R
\name{deltaSNPindex_plot}
\alias{deltaSNPindex_plot}
\title{delta(SNP-index) Wrapper Function}
\usage{
deltaSNPindex_plot(
  vcf.list,
  wtBulk,
  mBulk,
  variants = "SNP",
  min.SNPindex = 0.3,
  max.SNPindex = 0.9,
  min.DP = 50,
  max.DP = 200,
  min.GQ = 99,
  chrID,
  chr,
  windowSize = 1e+06,
  windowStep = 10000,
  filename = paste0("plot_deltaSNPindex_ch", chr),
  path = getwd(),
  dpi,
  width = 7.5,
  height = 5,
  units = "in"
)
}
\arguments{
\item{vcf.list}{object containing meta information and vcf data frame}

\item{wtBulk}{Wild-Type pool}

\item{mBulk}{Mutant pool}

\item{variants}{variants to be considered. Default is "SNP" (allowed: "SNP" or "all")}

\item{min.SNPindex}{min value allowed for the SNP index (default=0.3)}

\item{max.SNPindex}{max value allowed for the SNP index (default=0.9)}

\item{min.DP}{min value allowed for the read depth (default=50)}

\item{max.DP}{max value allowed for the read depth (default=200)}

\item{min.GQ}{min value allowed for the genotype quality (default=99)}

\item{chrID}{chromosome ID of interest}

\item{chr}{chromosome name printed on the plot}

\item{windowSize}{window size (default=1000000)}

\item{windowStep}{window step (default=10000)}

\item{filename}{file name under which the file will be saved (default="plot_deltaSNPindex_chX")}

\item{path}{path where the file will be saved (default=current working directory)}

\item{dpi}{resolution value. If no value is given, plots will be generated but not saved}

\item{width}{width value (default=7.5)}

\item{height}{height value (default=5)}

\item{units}{size units (default="in")}
}
\description{
This wrapper function is used to fully run the delta(SNP-index) method, by calling all the funcitons involved in plotting the delta(SNP-index) values from each bulk against the mid position of the corresponding window of a specific chromosome. 

\deqn{mean_M_SNPindex - mean_WT_SNPindex.WT}
}
\details{
Wrapper function that sequentially calls the required functions involved in generating the delta(SNP-index) plot:
calc_SNPindex(), filter_SNPindex(), extract_chrIDs(), slidingWindow(), calc_deltaSNPindex() and plot_deltaSNPindex(). 
The resulting plot will show (for each bulk) mean SNP-index values against the mid position of the corresponding window of a specific chromosome.
}
\examples{
## Use default values WITHOUT saving the plot
deltaSNPindex_plot(vcf.list=vcf_list, 
              wtBulk="pool_S3781_minus", 
              mBulk="pool_S3781_plus", 
              chrID="SL4.0ch03", 
              chr=3)
## OR use default values AND save the plot
deltaSNPindex_plot(vcf.list=vcf_list, 
              wtBulk="pool_S3781_minus", 
              mBulk="pool_S3781_plus", 
              chrID="SL4.0ch03", 
              chr=3, 
              dpi=1200)
#OR customise default parameters
deltaSNPindex_plot(vcf.list=vcf_list, 
              wtBulk="pool_S3781_minus", 
              mBulk="pool_S3781_plus", 
              variants="all",
              min.SNPindex=0.25, 
              max.SNPindex=0.8, 
              min.DP=60, 
              max.DP=250, 
              min.GQ=98,
              chrID="SL4.0ch03", 
              chr=3,
              windowSize=2000000, 
              windowStep=20000,
              filename="deltaSNPindex_chrom03", 
              path="Document/Plots", 
              dpi=1200,
              width=20, 
              height=12, 
              units="cm")   
}
