% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chrIDs.R
\name{extract_chrIDs}
\alias{extract_chrIDs}
\title{Extract Chromosome IDs}
\usage{
extract_chrIDs(meta)
}
\arguments{
\item{meta}{meta information}
}
\value{
Chromosome list
}
\description{
This function allows to extract the chromosome IDs, based on the meta information contained in the vcf file.
}
\examples{
chromList <- extract_chrIDs(meta=vcf_list$meta)
}
