% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_SNPindex.R
\name{filter_SNPindex}
\alias{filter_SNPindex}
\title{Filter SNP-index Variants}
\usage{
filter_SNPindex(
  vcf.df.SNPindex,
  min.SNPindex = 0.3,
  max.SNPindex = 0.9,
  min.DP = 50,
  max.DP = 200,
  min.GQ = 99
)
}
\arguments{
\item{vcf.df.SNPindex}{vcf dataframe}

\item{min.SNPindex}{min value allowed for the SNP index (default=0.3)}

\item{max.SNPindex}{max value allowed for the SNP index (default=0.9)}

\item{min.DP}{min value allowed for the read depth (default=50)}

\item{max.DP}{max value allowed for the read depth (default=200)}

\item{min.GQ}{min value allowed for the genotype quality (default=99)}
}
\value{
data frame containing variant and SNP-index information on variants that passed the filtering process.
}
\description{
This function filters out variants in the data frame returned by the calc_SNPindex() function 
which  do not  fall  between minimum DP, maximum DP, minimum SNP-index, maximum SNP-index, and minimum GQ for both bulks (wild-type and mutant).
}
\details{
Variants from the data frame returned by the calc_SNPindex() function 
are filtered out based on the following criteria:
minimum DP, maximum DP, minimum SNP-index, maximum SNP-index, and minimum GQ. 

Variants must pass these filters in both bulks to be retained.
}
\examples{
## Default parameters
vcf_df_SNPindex_filt <- filter_SNPindex(vcf.df.SNPindex=vcf_df_SNPindex)
## Custom parameters
vcf_df_SNPindex_filt <- filter_SNPindex(vcf.df.SNPindex=vcf_df_SNPindex, 
                                        min.SNPindex=0.25, 
                                        max.SNPindex=0.8, 
                                        min.DP=60, 
                                        max.DP=250, 
                                        min.GQ=98)
}
