% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_SNPratio.R
\name{filter_SNPratio}
\alias{filter_SNPratio}
\title{Filter SNP-index Variants}
\usage{
filter_SNPratio(vcf.df.SNPratio, min.SNPratio = 0.1, min.DP = 50, max.DP = 200)
}
\arguments{
\item{vcf.df.SNPratio}{vcf dataframe}

\item{min.SNPratio}{min value allowed for the SNP index (default=0.1)}

\item{min.DP}{min value allowed for the read depth (default=50)}

\item{max.DP}{max value allowed for the read depth (default=200)}
}
\value{
Data frame containing filtered variant information.
}
\description{
This functions allows to filter out bulk variants stored inside the data frame (returned by the calc_SNPratio() function) which do not fall between the 
minimumSNP-index, minimum DP and maximum DP.
}
\details{
Variants in the data frame returned by calc_SNPratio() with SNP-ratio values less below the minimum SNP-ratio value, as well as variants which do not fall between the given (or default) minimum and maximum DP values in both bulks, are discarded and removed from the final data frame returned by the function.
}
\examples{
## Default parameters
vcf_df_SNPratio_filt <- filter_SNPratio(vcf.df.SNPratio=vcf_df_SNPratio)
## Custom parameters
vcf_df_SNPratio_filt <- filter_SNPratio(vcf.df.SNPratio=vcf_df_SNPratio, 
                                        min.SNPratio=0.3, 
                                        min.DP=60, 
                                        max.DP=250)
}
