% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.r
\name{filter_variants}
\alias{filter_variants}
\title{This function allows to filter variants.}
\usage{
filter_variants(
  vcf.df.bulks,
  min_SNPindex = 0.3,
  max_SNPindex = 0.9,
  min_DP = 50,
  max_DP = 200,
  min_GQ = 99
)
}
\arguments{
\item{min_SNPindex}{min value allowed for the SNP index (default=0.3)}

\item{max_SNPindex}{max value allowed for the SNP index (default=0.9)}

\item{min_DP}{min value allowed for the read depth (default=50)}

\item{max_DP}{max value allowed for the read depth (default=200)}

\item{min_GQ}{min value allowed for the genotype quality (default=99)}
}
\value{
Filtered bulks (type: data frame)
}
\description{
This function allows to filter variants.
}
