% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deltaSNPindex.R
\name{plot_deltaSNPindex}
\alias{plot_deltaSNPindex}
\title{Plot delta(SNP index)}
\usage{
plot_deltaSNPindex(
  deltaSNPindex.windows,
  chr,
  filename = paste0("plot_deltaSNPindex_ch", chr),
  path = getwd(),
  dpi,
  width = 7.5,
  height = 5,
  units = "in"
)
}
\arguments{
\item{deltaSNPindex.windows}{filtered data frame (containing both bulks)}

\item{chr}{chrosome ID}

\item{filename}{file name under which the file will be saved (default="plot_deltaSNPindex_chX")}

\item{path}{path where the file will be saved (default=current working directory)}

\item{dpi}{resolution value. If no value is given, plots will be generated but not saved}

\item{width}{width value (default=7.5)}

\item{height}{height value (default=5)}

\item{units}{size units (default="in")}
}
\description{
This function allows plotting delta(SNP Index).
Whenever the DPI parameter gets specified, the plot will be saved in .TIF format with the specified resolution (dpi value).

\strong{Note:} \emph{for journal publications the preferred format is .TIF, and the minimum resolution is of 600 dpi or 1200 dpi. 
Always refer to journal-specific guidelines.}
}
\examples{
\strong{Plot SNP-index}
plot_deltaSNPindex(vcf.df.bulks=vcf_df_window_delta,
                   Chrom=3)
\strong{Save plot (specified resolution - tiff format)}
plot_deltaSNPindex(vcf.df.bulks=vcf_df_window_delta,
                   Chrom=3,
                   DPI=1200)
\strong{Save plot with specified width and height. The default unit is set to inches (Units="in")}
plot_deltaSNPindex(vcf.df.bulks=vcf_df_window_delta,
                   Chrom=3,
                   DPI=1200,
                   Height=9,
}
