% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SIMPLEratio.R
\name{plot_SIMPLEratio}
\alias{plot_SIMPLEratio}
\title{Plot SIMPLE-ratio}
\usage{
plot_SIMPLEratio(vcf.df.SIMPLEratio, min_SIMPLEratio, Chrom, DPI)
}
\arguments{
\item{vcf.df.SIMPLEratio}{SIMPLE-ratio data frame (containing both bulks)}

\item{min_SIMPLEratio}{minimum SIMPLE-ratio filtering value considered}

\item{Chrom}{chromosome ID}

\item{DPI}{resolution value. If no value is given, plots will be generated but not saved.}
}
\description{
This function allows plotting the SIMPLE-ratio (Wachsman et al., 2017).\n
The SIMPLE-ratio gets filtered by the minimum SIMPLE-ratio value (set by the user), before performing LOESS smoothing.\n
The plot gets generated based on the chromosome location of each SNP against the LOESS-fitted SIMPLE-ratio.\n
Whenever the DPI parameter gets specified, the plot will be saved in .TIF format with the specified resolution (dpi value).\n\n
Note: for journal publications the preferred format is .TIF, and the minimum resolution is of 600 dpi or 1200 dpi. Always refer to journal-specific guidelines.
}
\examples{
Plot SIMPLE-ratio for a specific chromosome\n
plot_SIMPLEratio(vcf.df.SIMPLEratio=vcf_df_SIMPLEratio, 
min_SIMPLEratio=0.1, 
Chrom=7)\n
Save plot (specified resolution - tiff format)\n
plot_SIMPLEratio(vcf.df.SIMPLEratio=vcf_df_SIMPLEratio, 
min_SIMPLEratio=0.1, 
Chrom=7,
DPI=1200)
}
