% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SNPratio.R
\name{plot_SNPratio}
\alias{plot_SNPratio}
\title{Plot SNP-ratio}
\usage{
plot_SNPratio(
  vcf.df.SNPratio.filt,
  chrList,
  chrID,
  chr,
  min.SNPratio,
  degree = 2,
  span = 0.07,
  filename = paste0("plot_SNPratio_ch", chr),
  path = getwd(),
  dpi,
  width = 7.5,
  height = 5,
  units = "in"
)
}
\arguments{
\item{vcf.df.SNPratio.filt}{filtered data frame (containing both bulks)}

\item{chrList}{list of chromosome IDs}

\item{chrID}{chromosome ID of interest}

\item{chr}{chromosome number to print on plot}

\item{min.SNPratio}{min SNP ratio threshold}

\item{degree}{LOESS smoothing degree (default=2)}

\item{span}{LOESS smoothing span (default=0.07)}

\item{filename}{file name under which the file will be saved (default="plot_SNPratio_chX")}

\item{path}{path where the file will be saved (default=current working directory)}

\item{dpi}{resolution value. If no value is given, plots will be generated but not saved}

\item{width}{width value (default=7.5)}

\item{height}{height value (default=5)}

\item{units}{size units (default="in")}
}
\description{
This function allows plotting the SNP-ratio (\emph{Wachsman et al., 2017}).
The SNP-ratio gets filtered by the minimum SNP-ratio value (set by the user), before performing LOESS smoothing.
The plot gets generated based on the chromosome location of each SNP against the LOESS-fitted SNP-ratio.
Whenever the DPI parameter gets specified, the plot will be saved in .TIF format with the specified resolution (dpi value).

\strong{Note:} \emph{for journal publications the preferred format is .TIF, and the minimum resolution is of 600 dpi or 1200 dpi. 
Always refer to journal-specific guidelines.}
}
\examples{
## Plot SNP-ratio for a specific chromosome
plot_SNPratio(vcf.df.SNPratio=vcf_df_SNPratio, 
              min_SNPratio=0.1, 
              Chrom=7)
## Save plot (specified resolution - tiff format)
plot_SNPratio(vcf.df.SNPratio=vcf_df_SNPratio, 
              min_SNPratio=0.1, 
              Chrom=7,
              DPI=1200)
## Save plot with specified width and height. The default unit is set to inches (Units="in")
plot_SNPratio(vcf.df.SNPratio=vcf_df_SNPratio, 
              min_SNPratio=0.1, 
              Chrom=7,
              DPI=1200)
              Height=9,
              Width=6)
}
\references{
Wachsman et al., 2017
}
