% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBSA_vcf.R
\name{readBSA_vcf}
\alias{readBSA_vcf}
\title{Read vcf file}
\usage{
readBSA_vcf(file)
}
\arguments{
\item{file}{vcf file}
}
\value{
a list containing meta information
}
\description{
This function allows to read a vcf file, converting it into a data frame. 
Format fields get separated as follows: AD, DP, GQ, GT. 

Allelic depths (AD) get split into reference and alterate AD values ("AD_ref" and "AD_alt").
}
\examples{
## Read vcf file
vcf_list <- readBSA_vcf("dataset1_pools.vcf")
}
