% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_SNPindex.R
\name{shiny_SNPindex}
\alias{shiny_SNPindex}
\title{SNP-index Wrapper Function}
\usage{
shiny_SNPindex(
  vcf.list,
  wtBulk,
  mBulk,
  variants,
  min.SNPindex,
  max.SNPindex,
  min.DP,
  max.DP,
  min.GQ,
  chrID,
  chr,
  windowSize,
  windowStep,
  bulk,
  ranges
)
}
\arguments{
\item{vcf.list}{object containing meta information and vcf data frame}

\item{wtBulk}{Wild-Type pool}

\item{mBulk}{Mutant pool}

\item{variants}{variants to be considered. Default is "SNP" (allowed: "SNP" or "all")}

\item{min.SNPindex}{min value allowed for the SNP index (default=0.3)}

\item{max.SNPindex}{max value allowed for the SNP index (default=0.9)}

\item{min.DP}{min value allowed for the read depth (default=50)}

\item{max.DP}{max value allowed for the read depth (default=200)}

\item{min.GQ}{min value allowed for the genotype quality (default=99)}

\item{chrID}{chromosome ID of interest}

\item{chr}{chromosome name printed on the plot}

\item{windowSize}{window size (default=1000000)}

\item{windowStep}{window step (default=10000)}

\item{bulk}{bulk/bulks to take into consideration}

\item{ranges}{axes ranges (x,y)}
}
\description{
Please note that this function is not to be run manually.
}
