% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_SNPratio.R
\name{shiny_SNPratio}
\alias{shiny_SNPratio}
\title{Shiny SNP-ratio Wrapper Function}
\usage{
shiny_SNPratio(
  vcf.list,
  wtBulk,
  mBulk,
  variants,
  min.SNPratio,
  min.DP,
  max.DP,
  chrID,
  chr,
  degree,
  span,
  ranges
)
}
\arguments{
\item{vcf.list}{object containing meta information and vcf data frame}

\item{wtBulk}{Wild-Type pool}

\item{mBulk}{Mutant pool}

\item{variants}{variants to be considered. Default is "SNP" (allowed: "SNP" or "all")}

\item{min.SNPratio}{min value allowed for the SNP index (default=0.3)}

\item{min.DP}{min value allowed for the read depth (default=50)}

\item{max.DP}{max value allowed for the read depth (default=200)}

\item{chrID}{chromosome ID of interest}

\item{chr}{chromosome name printed on the plot}

\item{degree}{LOESS smoothing degree (default=2)}

\item{span}{LOESS smoothing span (default=0.07)}

\item{ranges}{axes ranges (x,y)}
}
\description{
Please note that this function is not to be run manually.
Wrapper function for SNP-ratio plots specific to the R-shiny application. 
Calls all functions involved in plotting the SNP-ratio values across chromosome positions.
}
\details{
Wrapper function for generting SNP-ratio plots inside the R-Shiny application.
The involved functions are the following (in oder): 
calc_SNPratio(), filter_SNPratio(), extract_chrIDs() and shinyPlot_SNPratio().

This function is not intended to be run by the user, as it was specifically created to be used by the BSAvis_shiny() function.
}
