% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlot_deltaSNPindex.R
\name{shinyPlot_deltaSNPindex}
\alias{shinyPlot_deltaSNPindex}
\title{Plot delta(SNP index)}
\usage{
shinyPlot_deltaSNPindex(vcf.df.window.delta, chr, ranges)
}
\arguments{
\item{vcf.df.window.delta}{filtered data frame (containing both bulks)}

\item{chr}{chrosome ID}

\item{ranges}{axes ranges (x,y)}
}
\description{
Please note that this function is not to be run manually.
}
