% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlot_SNPratio.R
\name{shinyPlot_SNPratio}
\alias{shinyPlot_SNPratio}
\title{Shiny: Plot SNP-ratio}
\usage{
shinyPlot_SNPratio(
  vcf.df.SNPratio.filt,
  chrList,
  chrID,
  chr,
  min.SNPratio,
  degree = 2,
  span = 0.07,
  ranges
)
}
\arguments{
\item{vcf.df.SNPratio.filt}{filtered data frame (containing both bulks)}

\item{chrList}{list of chromosome IDs}

\item{chrID}{chromosome ID of interest}

\item{chr}{chromosome number to print on plot}

\item{min.SNPratio}{min SNP ratio threshold}

\item{degree}{LOESS smoothing degree (default=2)}

\item{span}{LOESS smoothing span (default=0.07)}

\item{ranges}{axes ranges (x,y)}
}
\description{
Please note that this function is not to be run manually.
Creates plots specific to the R-shiny application, showing SNP-ratio values against the position of a specific chromosome.
}
\details{
This function is a variant of the plot_SNPratio() function created to meet the needs of the R-shiny application.
The difference with the original is that it includes a "ranges" argument that allows setting the limits of the x and y-axis, to show a zoomed area of the SNP-ratio plot when selected. 
If no area is selected, the x and y axis limits will be set to null, showing the entire plot.

This function does not include any arguments related to plot saving since this functionality is linked to a saving button found inside the BSAvis R-Shiny application.
}
