% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPindex.R
\name{SNPindex}
\alias{SNPindex}
\title{SNP-index Wrapper Function}
\usage{
SNPindex(
  vcf.list,
  wtBulk,
  mBulk,
  variants = "SNP",
  min.SNPindex = 0.3,
  max.SNPindex = 0.9,
  min.DP = 50,
  max.DP = 200,
  min.GQ = 99,
  chrID,
  chr,
  windowSize = 1e+06,
  windowStep = 10000,
  filename = paste0("plot_SNPindex_ch", chr),
  path = getwd(),
  dpi,
  width = 7.5,
  height = 5,
  units = "in"
)
}
\arguments{
\item{vcf.list}{object containing meta information and vcf data frame}

\item{wtBulk}{Wild-Type pool}

\item{mBulk}{Mutant pool}

\item{variants}{variants to be considered. Default is "SNP" (allowed: "SNP" or "all")}

\item{min.SNPindex}{min value allowed for the SNP index (default=0.3)}

\item{max.SNPindex}{max value allowed for the SNP index (default=0.9)}

\item{min.DP}{min value allowed for the read depth (default=50)}

\item{max.DP}{max value allowed for the read depth (default=200)}

\item{min.GQ}{min value allowed for the genotype quality (default=99)}

\item{chrID}{chromosome ID of interest}

\item{chr}{chromosome name printed on the plot}

\item{windowSize}{window size (default=1000000)}

\item{windowStep}{window step (default=10000)}

\item{filename}{file name under which the file will be saved (default="plot_SNPindex_chX")}

\item{path}{path where the file will be saved (default=current working directory)}

\item{dpi}{resolution value. If no value is given, plots will be generated but not saved}

\item{width}{width value (default=7.5)}

\item{height}{height value (default=5)}

\item{units}{size units (default="in")}
}
\description{
This function is used to fully run the SNP-index method.

\deqn{SNPindex=AD_alt/(AD_ref + AD_alt)}
}
\examples{
vcf_df_window_delta <- calc_deltaSNPindex(SNPindex.windows=vcf_df_window)
}
