% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_deltaSNPindex.R
\name{calc_deltaSNPindex}
\alias{calc_deltaSNPindex}
\title{Calculate delta(SNP index)}
\usage{
calc_deltaSNPindex(SNPindex.windows)
}
\arguments{
\item{SNPindex.windows}{data frame obtained after running the slidingWindow()function}
}
\value{
data frame containing the delta(SNP-index) means values
}
\description{
This function is used to calculate the delta(SNP index) using the calculated mean of the mutant (M) and wild-type (WT) indices, as follows:
\deqn{mean_M_SNPindex - mean_WT_SNPindex.WT}
}
\examples{
deltaSNPindex_windows <- calc_deltaSNPindex(SNPindex.windows=SNPindex_windows) 
}
