% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_deltaSNPindex.r
\name{calc_deltaSNPindex}
\alias{calc_deltaSNPindex}
\title{This function is used to calculate the delta(SNP index) using the calculated mean of the M and WT indices, as follows:
\deqn{mean_SNPindex.M - mean_SNPindex.WT)}.}
\usage{
calc_deltaSNPindex(vcf.df.window)
}
\arguments{
\item{vcf.df.window}{data frame obtained after running the slidingWindow() function}
}
\value{
data frame containing the delta(SNP index) values
}
\description{
This function is used to calculate the delta(SNP index) using the calculated mean of the M and WT indices, as follows:
\deqn{mean_SNPindex.M - mean_SNPindex.WT)}.
}
\examples{
calc_deltaSNPindex()
}
