% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SNPIndex.r
\name{calc_SNPindex}
\alias{calc_SNPindex}
\title{This function is used to sparate (filter) the two bulks, and calculate their (SNP index) as follows:
\deqn{SNPindex=as.numeric(AD_alt)/(as.numeric(AD_ref) + as.numeric(AD_alt)}. 
Bulks are finally joined together in a single data frame.}
\usage{
calc_SNPindex(vcf.df, WTbulk, Mbulk)
}
\arguments{
\item{vcf.df}{Data frame of the vcf file}

\item{WTbulk}{Wild-Type pool}

\item{Mbulk}{Mutant pool}
}
\value{
A dataframe with the joint bulks
}
\description{
This function is used to sparate (filter) the two bulks, and calculate their (SNP index) as follows:
\deqn{SNPindex=as.numeric(AD_alt)/(as.numeric(AD_ref) + as.numeric(AD_alt)}. 
Bulks are finally joined together in a single data frame.
}
\examples{
calc_SNPindex()
}
