% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SNPIndex.R
\name{calc_SNPindex}
\alias{calc_SNPindex}
\title{Calculate SNP-index}
\usage{
calc_SNPindex(vcf.df, wtBulk, mBulk, variants = "SNP")
}
\arguments{
\item{vcf.df}{Data frame of the vcf file}

\item{wtBulk}{Wild-Type pool}

\item{mBulk}{Mutant pool}

\item{variants}{variants to be considered. Default is "SNP" (allowed: "SNP" or "all")}
}
\value{
A dataframe with the joint bulks
}
\description{
This function is used to sparate (filter) the two bulks, and calculate their (SNP index) as follows:
\deqn{SNPindex=AD_alt/(AD_ref + AD_alt)} 
Bulks get joined together in a single data frame.

Note that the user can select specific variants to consider, 
by setting the "variants" parameter to "SNP" (default) or "all" (InDels+SNPs).
}
\examples{
## Calculate SNP-index for both bulks (only SNPs will be considered)
vcf_df_SNPindex <- calc_SNPindex(vcf.df=vcf_list$df, 
                                 wtBulk="pool_S3781_minus", 
                                 mBulk="pool_S3781_plus") 

## Calculate SNP-index considering both InDels and SNPs
vcf_df_SNPindex <- calc_SNPindex(vcf.df=vcf_list$df, 
                                 wtBulk="pool_S3781_minus", 
                                 mBulk="pool_S3781_plus",
                                 variants="all") 
}
