% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SNPratio.R
\name{calc_SNPratio}
\alias{calc_SNPratio}
\title{Calculate SNP-ratio}
\usage{
calc_SNPratio(vcf.df, wtBulk, mBulk, variants = "SNP")
}
\arguments{
\item{vcf.df}{Data frame of the vcf file}

\item{wtBulk}{Wild-Type pool}

\item{mBulk}{Mutant pool}

\item{variants}{variants to be considered. Default is "SNP" (allowed: "SNP" or "all")}
}
\value{
data frame (SNP-ratio)
}
\description{
This function allows to calculate the SNP-ratio between two bulks (M/WT), as follows:
\deqn{(AD_ref.WT/(AD_ref.WT + AD_alt.WT)) - (AD_ref.M/(AD_ref.M + AD_alt.M))}
Note that the user can select specific variants to consider, by setting the "Variants" parameter to "SNP" (default) or "all" (respectively referring to SNPs or InDels+SNPs).
}
\examples{
## Calculate SNP-index for both bulks (only SNPs will be considered)
vcf_df_SNPratio <- calc_SNPratio(vcf.df=vcf_list$df, 
                                 wtBulk="pool_S3781_minus", 
                                 mBulk="pool_S3781_plus") 
## Calculate SNP-index considering both InDels and SNPs
vcf_df_SNPratio <- calc_SNPratio(vcf.df=vcf_list$df, 
                                 wtBulk="pool_S3781_minus", 
                                 mBulk="pool_S3781_plus", 
                                 variants="all")
}
\references{
Wachsman et al., 2017
}
