% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chrIDs.R
\name{extract_chrIDs}
\alias{extract_chrIDs}
\title{Extract Chromosome IDs}
\usage{
extract_chrIDs(meta)
}
\arguments{
\item{meta}{meta information contained in the first element of the list returned by the readBSA_vcf() function}
}
\value{
Character vector containing a list of chromosome IDs.
}
\description{
This function allows to extract the chromosome IDs, based on the meta information contained in the VCF file.
}
\details{
The meta information from the VCF file (which is stored in the first element of the list generated byreadBSA_vcf())is taken as input. 

The lines in the meta information which contain a sequence of key words (including ID and length) that make them unique from others,  
are extracted into a character vector. 

For each element of that vector, the characters before and after the chromosome ID are removed, resulting in each element of the vector containing only the characters corresponding to a specific chromosome ID.
The final character vector contains all chromosomes IDs in the way they are named and ordered within the VCF file.
}
\examples{
chromList <- extract_chrIDs(meta=vcf_list$meta)
}
