% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chromLength.r
\name{extract_chromLength}
\alias{extract_chromLength}
\title{This function allows to extract the chromosome length of a specific chromosome, using the meta information contained in the vcf file.
this function extract chromosome length}
\usage{
extract_chromLength(chromID, meta)
}
\arguments{
\item{chromID}{chromosome ID}

\item{vcf}{}
}
\description{
This function allows to extract the chromosome length of a specific chromosome, using the meta information contained in the vcf file.
this function extract chromosome length
}
\examples{
extract_chromLength()
}
