% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_SNPindex.R
\name{filter_SNPindex}
\alias{filter_SNPindex}
\title{Filter SNP-index Variants}
\usage{
filter_SNPindex(
  vcf.df.SNPindex,
  min.SNPindex = 0.3,
  max.SNPindex = 0.9,
  min.DP = 50,
  max.DP = 200,
  min.GQ = 99
)
}
\arguments{
\item{vcf.df.SNPindex}{vcf dataframe}

\item{min.SNPindex}{min value allowed for the SNP index (default=0.3)}

\item{max.SNPindex}{max value allowed for the SNP index (default=0.9)}

\item{min.DP}{min value allowed for the read depth (default=50)}

\item{max.DP}{max value allowed for the read depth (default=200)}

\item{min.GQ}{min value allowed for the genotype quality (default=99)}
}
\value{
Filtered bulks for the SNP-index method (type: data frame)
}
\description{
This function allows to filter variants for the SNP-index method.
}
\examples{
## Default parameters
vcf_df_SNPindex_filt <- filter_SNPindex(vcf.df.SNPindex=vcf_df_SNPindex)
## Custom parameters
vcf_df_SNPindex_filt <- filter_SNPindex(vcf.df.SNPindex=vcf_df_SNPindex, 
                                        min.SNPindex=0.25, 
                                        max.SNPindex=0.8, 
                                        min.DP=60, 
                                        max.DP=250, 
                                        min.GQ=98)
}
