% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deltaSNPindex.r
\name{plot_deltaSNPindex}
\alias{plot_deltaSNPindex}
\title{This function allows plotting delta(SNP Index).}
\usage{
plot_deltaSNPindex(vcf.df.bulks.filt, Chrom)
}
\arguments{
\item{vcf.df.bulks.filt}{filtered data frame (containing both bulks)}

\item{Chrom}{chromosome ID}
}
\description{
This function allows plotting delta(SNP Index).
}
\examples{
plot_deltaSNPindex()
}
