% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deltaSNPindex.R
\name{plot_deltaSNPindex}
\alias{plot_deltaSNPindex}
\title{Plot delta(SNP index)}
\usage{
plot_deltaSNPindex(
  deltaSNPindex.windows,
  chr,
  filename = paste0("plot_deltaSNPindex_ch", chr),
  path = getwd(),
  dpi,
  width = 7.5,
  height = 5,
  units = "in"
)
}
\arguments{
\item{deltaSNPindex.windows}{filtered data frame (containing both bulks)}

\item{chr}{chrosome ID}

\item{filename}{file name under which the file will be saved (default="plot_deltaSNPindex_chX")}

\item{path}{path where the file will be saved (default=current working directory)}

\item{dpi}{resolution value. If no value is given, plots will be generated but not saved}

\item{width}{width value (default=7.5)}

\item{height}{height value (default=5)}

\item{units}{size units (default="in")}
}
\description{
This function allows plotting the delta(SNP-index) values against the mid position of the corresponding window of a specific chromosome. 
By setting the dpi parameter (resolution), the plot will be automatically saved in .TIF format.

\strong{Note:} \emph{for journal publications the preferred format is .TIF, and the minimum resolution is of 600 dpi or 1200 dpi. 
Always refer to journal-specific guidelines.}
}
\details{
∆(SNP-index) values are plotted against window mid position in a ggplot2line plot. If the dpi argument is not passed, the plot will be shown in the plot panel; however, if a value is added to the dpi argument, the graph will be saved in TIFF format. 
Additionally, the name with which to save the file, the directory, the height and width of the plot and their units can be specified in the arguments. 
The plot will be saved with default values if different ones are not specified.
}
\examples{
## Use default values WITHOUT saving the plot
plot_deltaSNPindex(deltaSNPindex.windows=deltaSNPindex_windows, 
                   chr=3)
## OR use default values AND save the plot
plot_deltaSNPindex(deltaSNPindex.windows=deltaSNPindex_windows, 
                   chr=3,
                   dpi=1200)
## OR customise default parameters
plot_deltaSNPindex(deltaSNPindex.windows=deltaSNPindex_windows, 
                   chr=3,
                   filename="deltaSNPindex_ch3", 
                   path="Document/Plots", 
                   dpi=1200, 
                   width=20, 
                   height=12, 
                   units="cm")
}
