% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SNPindex.r
\name{plot_SNPindex}
\alias{plot_SNPindex}
\title{This function allows plotting the (SNP Index).}
\usage{
plot_SNPindex(vcf.df.bulks.filt, Chrom)
}
\arguments{
\item{vcf.df.bulks.filt}{filtered data frame (containing both bulks)}

\item{Chrom}{chromosome ID}
}
\description{
This function allows plotting the (SNP Index).
}
\examples{
plot_SNPindex()
}
