% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SNPindex.R
\name{plot_SNPindex}
\alias{plot_SNPindex}
\title{Plot SNP-indices}
\usage{
plot_SNPindex(
  SNPindex.windows,
  chr,
  filename = paste0("plot_SNPindex_ch", chr),
  path = getwd(),
  dpi,
  width = 7.5,
  height = 5,
  units = "in"
)
}
\arguments{
\item{SNPindex.windows}{filtered data frame (containing both bulks)}

\item{chr}{chrosome ID}

\item{filename}{file name under which the file will be saved (default="plot_SNPindex_chX")}

\item{path}{path where the file will be saved (default=current working directory)}

\item{dpi}{resolution value. If no value is given, plots will be generated but not saved}

\item{width}{width value (default=7.5)}

\item{height}{height value (default=5)}

\item{units}{size units (default="in")}
}
\description{
This function allows plotting the mean SNP-index values from each bulk against the mid position of the corresponding window of a specific chromosome. 
By setting the dpi parameter (resolution), the plot will be automatically saved in .TIF format.

\strong{Note:} \emph{for journal publications the preferred format is .TIF, and the minimum resolution is of 600 dpi or 1200 dpi. 
Always refer to journal-specific guidelines.}
}
\details{
The mean SNP-index from each bulk is plotted in a different colour against the corresponding window mid position in a ggplot2 line plot.

If the dpi argument is not passed, the plot will be shown in the plot panel; however, if a value is added to the dpi argument, the graph will be saved in TIFF format. 
Additionally, the name with which to save the file, the directory, the height and width of the plot and their units can be specified in the arguments. 

The plot will be saved with default values if different ones are not specified.
}
\examples{
## OR use default values AND save the plot
plot_SNPindex(SNPindex.windows=SNPindex_windows, 
              chr=3,
              dpi=1200) 
## OR customise default parameters
plot_SNPindex(SNPindex.windows=SNPindex_windows, 
              chr=3, 
              filename="SNPindex_ch3", 
              path="Document/Plots", 
              dpi=1200, 
              width=20, 
              height=12, 
              units="cm") 
}
