% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SNPratio.R
\name{plot_SNPratio}
\alias{plot_SNPratio}
\title{Plot SNP-ratio}
\usage{
plot_SNPratio(
  vcf.df.SNPratio,
  min_SNPratio = 0.1,
  Degree = 2,
  Span = 0.3,
  Chrom,
  DPI,
  Width = 7.5,
  Height = 5,
  Units = "in"
)
}
\arguments{
\item{vcf.df.SNPratio}{SNP-ratio data frame (containing both bulks)}

\item{min_SNPratio}{minimum SNP-ratio value to consider (default=0.1)}

\item{Degree}{LOESS smoothing degree (default=2)}

\item{Span}{LOESS smoothing span (default=0.3)}

\item{Chrom}{chromosome ID}

\item{DPI}{resolution value. If no value is given, plots will be generated but not saved}

\item{Width}{width value (default=7.5)}

\item{Height}{height value (default=5)}

\item{Units}{size units (default="in")}
}
\description{
This function allows plotting the SNP-ratio (\emph{Wachsman et al., 2017}).
The SNP-ratio gets filtered by the minimum SNP-ratio value (set by the user), before performing LOESS smoothing.
The plot gets generated based on the chromosome location of each SNP against the LOESS-fitted SNP-ratio.
Whenever the DPI parameter gets specified, the plot will be saved in .TIF format with the specified resolution (dpi value).

\strong{Note:} \emph{for journal publications the preferred format is .TIF, and the minimum resolution is of 600 dpi or 1200 dpi. 
Always refer to journal-specific guidelines.}
}
\examples{
\strong{Plot SNP-ratio for a specific chromosome}
plot_SNPratio(vcf.df.SNPratio=vcf_df_SNPratio, 
              min_SNPratio=0.1, 
              Chrom=7)
\strong{Save plot (specified resolution - tiff format)}
plot_SNPratio(vcf.df.SNPratio=vcf_df_SNPratio, 
              min_SNPratio=0.1, 
              Chrom=7,
              DPI=1200)
\strong{Save plot with specified width and height. The default unit is set to inches (Units="in")}
plot_SNPratio(vcf.df.SNPratio=vcf_df_SNPratio, 
              min_SNPratio=0.1, 
              Chrom=7,
              DPI=1200)
              Height=9,
              Width=6)
}
\references{
Wachsman et al., 2017
}
