% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_SNPratio.R
\name{shiny_SNPratio}
\alias{shiny_SNPratio}
\title{Shiny SNP-ratio Wrapper Function}
\usage{
shiny_SNPratio(
  vcf.list,
  wtBulk,
  mBulk,
  variants,
  min.SNPratio,
  min.DP,
  max.DP,
  chrID,
  chr,
  degree,
  span,
  ranges
)
}
\arguments{
\item{vcf.list}{object containing meta information and vcf data frame}

\item{wtBulk}{Wild-Type pool}

\item{mBulk}{Mutant pool}

\item{variants}{variants to be considered. Default is "SNP" (allowed: "SNP" or "all")}

\item{min.SNPratio}{min value allowed for the SNP index (default=0.3)}

\item{min.DP}{min value allowed for the read depth (default=50)}

\item{max.DP}{max value allowed for the read depth (default=200)}

\item{chrID}{chromosome ID of interest}

\item{chr}{chromosome name printed on the plot}

\item{degree}{LOESS smoothing degree (default=2)}

\item{span}{LOESS smoothing span (default=0.07)}

\item{ranges}{axes ranges (x,y)}
}
\description{
Please note that this function is not to be run manually.
}
