% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlot_SNPindex.R
\name{shinyPlot_SNPindex}
\alias{shinyPlot_SNPindex}
\title{Shiny: Plot SNP-indices}
\usage{
shinyPlot_SNPindex(vcf.df.window.SNPindex, chr, bulk, ranges)
}
\arguments{
\item{chr}{chrosome ID}

\item{bulk}{bulk/bulks to take into consideration when plotting}

\item{ranges}{axes ranges (x,y)}

\item{SNPindex.windows}{filtered data frame (containing both bulks)}
}
\description{
Please note that this function is not to be run manually.
}
