% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingWindow.R
\name{slidingWindow}
\alias{slidingWindow}
\title{Sliding Windows}
\usage{
slidingWindow(
  meta,
  chrList,
  chrID,
  windowSize = 1e+06,
  windowStep = 10000,
  vcf.df.SNPindex.filt
)
}
\arguments{
\item{meta}{meta information stored inside the vcf file}

\item{chrList}{list of chromosome IDs}

\item{chrID}{chromosome ID of interest}

\item{windowSize}{window size (default=1000000)}

\item{windowStep}{window step (default=10000)}

\item{vcf.df.SNPindex.filt}{filtered SNP-index data frame}
}
\value{
SNP-index windows
}
\description{
This function allows to calculate the sliding window based on the chromosome length of the desired chromosome to analyse. 
A data frame containing the mean SNP-index values of both bulks gets generated.
}
\examples{
## Default parameters
SNPindex_windows <- slidingWindow(meta=vcf_list$meta, 
                                  chrList=chromList, 
                                  chrID="SL4.0ch03",  
                                  vcf.df.SNPindex.filt=vcf_df_SNPindex_filt)
## Custom parameters
SNPindex_windows <- slidingWindow(meta=vcf_list$meta, 
                                  chrList=chromList, 
                                  chrID="SL4.0ch03", 
                                  windowSize=2000000, 
                                  windowStep=20000, 
                                  vcf.df.SNPindex.filt=vcf_df_SNPindex_filt)
}
