% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingWindow.r
\name{slidingWindow}
\alias{slidingWindow}
\title{This function allows to calculate the sliding window and the mean values of both bulks.}
\usage{
slidingWindow(Chrom, chromLength, windowSize, windowStep, df)
}
\arguments{
\item{Chrom}{chromosome ID}

\item{chromLength}{chromosome length}

\item{windowSize}{window size}

\item{windowStep}{window step}

\item{df}{data frame}
}
\value{
windows
}
\description{
This function allows to calculate the sliding window and the mean values of both bulks.
}
\examples{
slidingwindow()
}
